+++ keyring/examples/keyring.rs
@@ -6,15 +6,16 @@ const VAL: &str = "val";
 fn main() {
     env_logger::builder().is_test(true).init();
 
-    // set entry
+    // set entry secret
     let entry = Entry::from(KEY);
-    entry.set(VAL).unwrap();
+    entry.set_secret(VAL).unwrap();
 
-    // get entry
-    let val = entry.get().unwrap();
+    // get entry secret
+    let val = entry.get_secret().unwrap();
+    assert_eq!(entry.get_key(), KEY);
     assert_eq!(val, VAL);
 
     // delete entry
-    entry.delete().unwrap();
-    assert_eq!(entry.find().unwrap(), None);
+    entry.delete_secret().unwrap();
+    assert_eq!(entry.find_secret().unwrap(), None);
 }