+++ keyring/src/lib.rs
@@ -1,5 +1,5 @@
 pub use keyring::Error as KeyringError;
-use log::debug;
+use log::{debug, trace};
 use std::{ops::Deref, result};
 use thiserror::Error;
 
@@ -7,11 +7,13 @@ use thiserror::Error;
 pub enum Error {
     #[error("cannot get keyring entry {1}")]
     GetEntryError(#[source] KeyringError, String),
-    #[error("cannot get keyring secret for entry {1}")]
+    #[error("cannot get keyring entry secret for key {1}")]
     GetSecretError(#[source] KeyringError, String),
-    #[error("cannot set keyring secret for entry {1}")]
+    #[error("cannot find keyring entry secret for key {1}")]
+    FindSecretError(#[source] KeyringError, String),
+    #[error("cannot set keyring entry secret for key {1}")]
     SetSecretError(#[source] KeyringError, String),
-    #[error("cannot delete keyring secret for entry {1}")]
+    #[error("cannot delete keyring entry secret for key {1}")]
     DeleteSecretError(#[source] KeyringError, String),
 }
 
@@ -22,52 +24,59 @@ const KEYRING_SERVICE: &str = "pimalaya";
 /// Alias for the keyring entry key.
 pub type Key = String;
 
-/// Wrapper around [`keyring::Entry`] that holds a keyring entry key.
+/// Keyring entry. It is a simple wrapper around [`keyring::Entry`]
+/// that holds a keyring entry key.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct Entry(Key);
 
 impl Entry {
-    /// Create a new keyring [`Entry`] based on the given key.
-    pub fn new(key: impl ToString) -> Self {
-        Self(key.to_string())
-    }
-
     fn get_entry(&self) -> Result<keyring::Entry> {
         keyring::Entry::new(KEYRING_SERVICE, &self)
             .map_err(|err| Error::GetEntryError(err, self.0.clone()))
     }
 
-    /// Get the secret from the user's global keyring.
-    pub fn get(&self) -> Result<String> {
-        debug!("getting keyring secret for entry {:?}", self.0);
+    /// Create a new keyring entry with a keyring entry key.
+    pub fn new(key: impl ToString) -> Self {
+        Self(key.to_string())
+    }
+
+    /// Return the inner key of the keyring entry.
+    pub fn get_key(&self) -> &str {
+        debug!("getting keyring entry key");
+        trace!("keyring entry key: {:?}", self.0);
+        self.as_str()
+    }
+
+    /// Get the secret of the keyring entry.
+    pub fn get_secret(&self) -> Result<String> {
+        debug!("getting keyring entry secret for key {:?}", self.0);
         self.get_entry()?
             .get_password()
             .map_err(|err| Error::GetSecretError(err, self.0.clone()))
     }
 
-    /// Find the secret from the user's global keyring. Similar to
-    /// [`Entry::get`], except that it returns None in case the entry
-    /// is not found.
-    pub fn find(&self) -> Result<Option<String>> {
-        debug!("finding keyring secret for entry {:?}", self.0);
+    /// Find the secret of the keyring entry. Return None in case the
+    /// secret is not found.
+    pub fn find_secret(&self) -> Result<Option<String>> {
+        debug!("finding keyring entry secret for key {:?}", self.0);
         match self.get_entry()?.get_password() {
             Err(keyring::Error::NoEntry) => Ok(None),
-            Err(err) => Err(Error::GetSecretError(err, self.0.clone())),
+            Err(err) => Err(Error::FindSecretError(err, self.0.clone())),
             Ok(secret) => Ok(Some(secret)),
         }
     }
 
-    /// (Re)set the secret from the user's global keyring.
-    pub fn set(&self, secret: impl AsRef<str>) -> Result<()> {
-        debug!("setting keyring secret for entry {:?}", self.0);
+    /// (Re)set the secret of the keyring entry.
+    pub fn set_secret(&self, secret: impl AsRef<str>) -> Result<()> {
+        debug!("setting keyring entry secret for key {:?}", self.0);
         self.get_entry()?
             .set_password(secret.as_ref())
             .map_err(|err| Error::SetSecretError(err, self.0.clone()))
     }
 
-    /// Delete the secret from the user's global keyring.
-    pub fn delete(&self) -> Result<()> {
-        debug!("deleting keyring secret for entry {:?}", self.0);
+    /// Delete the secret of the keyring entry.
+    pub fn delete_secret(&self) -> Result<()> {
+        debug!("deleting keyring entry secret for key {:?}", self.0);
         self.get_entry()?
             .delete_password()
             .map_err(|err| Error::DeleteSecretError(err, self.0.clone()))
@@ -97,6 +106,12 @@ impl From<&String> for Entry {
 impl From<&str> for Entry {
     fn from(key: &str) -> Self {
         Self::new(key)
+    }
+}
+
+impl Into<String> for Entry {
+    fn into(self) -> String {
+        self.0
     }
 }